desktop-app/requirements.txt
flask
pyautogui
pillow
google-genai
